+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 ### Fixed
+- Fixed FlowClient ignoring custom SSL certificates [#3344](https://github.com/Flowminder/FlowKit/issues/3344)
 
 ### Removed